requirements.txt
streamlit>=1.30.0
opencv-python>=4.8.0
Pillow>=10.0.0
PyYAML>=6.0
aiohttp>=3.9.0
APScheduler>=3.10.0
requests>=2.31.0
numpy>=1.24.0
modelscope>=1.11.0
ultralytics>=8.1.0
